LSR-1122 Move test to services-v7 project
