+++ android/android-sdk/build.gradle
@@ -4,7 +4,4 @@ dependencies {
 	implementation group: 'com.squareup.okhttp', name: 'okhttp', version: '2.7.5'
 
 	implementation 'net.openid:appauth:0.7.0'
-
-	testImplementation 'org.json:json:20140107'
-	testImplementation 'junit:junit:4.12'
 }
