+++ android/services-v7/src/test/java/com/liferay/mobile/android/UnauthenticatedServiceTest.java
