+++ android/services-v7/src/test/resources/test.properties
@@ -1,0 +1,5 @@
+companyId=
+groupId=
+url=http://localhost:8080
+login=test@liferay.com
+password=test
